backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
pdfplumber==0.9.0
python-dotenv==1.0.0
Pillow==10.1.0 